requirements.txt
praw
langchain
python-dotenv
langchain_groq
streamlit
tenacity
langchain_community
-e .